move BaseTW type def to tw.ts; use NumericBin type instead of RangeEntry in numeric termsetting/handler code
